+++ common/src/main/java/fr/tathan/sky_aesthetics/client/skies/record/CloudSettings.java
@@ -3,7 +3,6 @@ package fr.tathan.sky_aesthetics.client.skies.record;
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
 import net.minecraft.world.phys.Vec3;
-import org.joml.Vector3f;
 
 import java.util.Optional;